distance display + consumed power display

+++ Core/Src/Driver/KT/Kt.h
@@ -16,13 +16,13 @@ private:
     kt::KtToLcdPayload mCurrentKtToLcdPayload;
     kt::LcdToKtPayload mCurrentLcdToKtPayload;
     bool mNeedDispatchPayload = false;
-    float mCurrentSpeed = 0.f;
+    double mCurrentSpeed = 0.f;
     std::uint16_t mMotorPower = 0;
     std::uint8_t mMotorTemperature = 0;
-    float mTmp = 0.f;
+    double mTmp = 0.f;
     kt::Config mConfig;
 
-    float wheelDiameterInch() const;
+    double wheelDiameterInch() const;
 };
